Hashmap implementation now uses array implementation under the hood, This is to reduce the amount of memory handling needed to be done which was already implemented in array implmentation

+++ include/utils/hashmap.h
@@ -13,4 +13,4 @@ extern StatusCode hashmap_AddEntry(StrIntHashmap *hashmap, String key,
 extern int64_t hashmap_FetchValue(StrIntHashmap *hashmap, String key);
 extern StatusCode hashmap_DeleteEntry(StrIntHashmap *hashmap,
                                       String key);
-extern uint64_t hashmap_GetLen(StrIntHashmap *hashmap);
+extern size_t hashmap_GetLen(StrIntHashmap *hashmap);
